refactor: storage.ts の型定義を分割整理
- types/analytics.ts を作成（DailyStat, SiteTime, AnalyticsData 等）
- types/vision.ts を作成（VisionSettings, DashboardPreset 等）
- types/index.ts を作成（統一エクスポート）
- storage.ts は互換性のため re-export を維持
- messageSchemas.ts と messages.ts の重複を Zod スキーマからの型推論に統一
